[feat] : integrates configuration and customization modules
