--- customization-module/config.json
@@ -1,11 +1,0 @@
-{
-    "font" : {
-        "size" : "20px",
-        "colour": "#000000"
-    },
-
-    "background" : {
-        "opacity": "100%",
-        "color": "#ffffff"
-    }
-}
