+++ integration-test/ConfigurationInterface.js
