+++ integration-test/WindowHandler.js
