+++ integration-test/index.html
@@ -5,7 +5,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Test App</title>
 </head>
-<body>
-    <p id="main-title">Hello, world!</p>
+<body id="content">
+
+    <!-- ADD EDITOR WRAPPER HERE -->
+    <h3>Used for testing purposes</h3>
+    <p>This too, by the way...</p>
+
 </body>
 </html>
